// src/domain/instructors/instructor.interface.ts
export interface Iinstructor {
    _id: string;
    firstName: string;
    lastName: string;
    avatar: string;
    phone: string;
    about: string[] ;
    speciality: string[];
    technologies: string[];
    experience: string;
    intExperience: string;
    studyAt: string[];
    workingAt: string[];
    workedAt: string[];
    inFulfilSince: string;
}